auto-complete: fill to the brim with options
ac-html is an option now, so delete its formula.

Most of these still depend on auto-complete, but js2 and probably others
can/will be able to use company-mode instead.  Since it's not clear to
the user which are, they should be added as options to company-mode as
well when they are compatible.